practica_1/requirements.txt
pandas
GitPython 
requests
json